Move new test logic back to original folder
