+++ test/amp/destination/agent_configs/otlp_config.json
