+++ test/amp/destination/agent_configs/prometheus_config.json
