+++ test/amp/destination/amp_test.go
@@ -8,10 +8,12 @@ package amp
 import (
 	"context"
 	"crypto/sha256"
+	_ "embed"
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"io"
+	"log"
 	"net/http"
 	"path/filepath"
 	"strconv"
@@ -22,6 +24,7 @@ import (
 	sigv4 "github.com/aws/aws-sdk-go-v2/aws/signer/v4"
 	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/cloudwatch/types"
+	"github.com/stretchr/testify/suite"
 
 	"github.com/aws/amazon-cloudwatch-agent-test/environment"
 	"github.com/aws/amazon-cloudwatch-agent-test/test/metric"
@@ -30,6 +33,70 @@ import (
 	"github.com/aws/amazon-cloudwatch-agent-test/test/test_runner"
 	"github.com/aws/amazon-cloudwatch-agent-test/util/common"
 )
+
+const (
+	// template prometheus query for getting average of 3 min
+	ampQueryTemplate          = "avg_over_time(%s%s[3m])"
+	ampHistogramQueryTemplate = "quantile_over_time(0.5, %s_bucket%s[3m])"
+)
+
+type Source int
+
+const (
+	SourcePrometheus Source = iota
+	SourceOtlp
+	SourceHost
+)
+
+// NOTE: this should match with append_dimensions under metrics in agent config
+var appendDims = map[string]string{
+	"d1": "foo",
+	"d2": "bar",
+}
+
+var awsConfig aws.Config
+var awsCreds aws.Credentials
+var metadata *environment.MetaData
+
+//go:embed resources/prometheus.yaml
+var prometheusConfig string
+
+//go:embed resources/prometheus_metrics
+var prometheusMetrics string
+
+var (
+	testRunners []*test_runner.TestRunner = []*test_runner.TestRunner{
+		{
+			TestRunner: &AmpTestRunner{
+				source: SourceHost,
+				config: "host_config.json",
+				name:   "host",
+			},
+		},
+		{
+			TestRunner: &AmpTestRunner{
+				source: SourcePrometheus,
+				config: "prometheus_config.json",
+				name:   "prometheus",
+			},
+		},
+		{
+			TestRunner: &AmpTestRunner{
+				source: SourceOtlp,
+				config: "otlp_config.json",
+				name:   "otlp",
+			},
+		},
+	}
+)
+
+func init() {
+	environment.RegisterEnvironmentMetaDataFlags()
+}
+
+func TestAmpTestSuite(t *testing.T) {
+	suite.Run(t, new(AmpTestSuite))
+}
 
 type AMPResponse struct {
 	Status string
@@ -44,23 +111,22 @@ type AMPDataResult struct {
 	Value  []interface{}
 }
 
-const (
-	namespace = "AMPDestinationTest"
-	// template prometheus query for getting average of 3 min
-	ampQueryTemplate = "avg_over_time(%s%s[3m])"
-)
-
-// NOTE: this should match with append_dimensions under metrics in agent config
-var append_dims = map[string]string{
-	"d1": "foo",
-	"d2": "bar",
+type AmpTestSuite struct {
+	suite.Suite
+	test_runner.TestSuite
 }
 
-var awsConfig aws.Config
-var awsCreds aws.Credentials
+func (suite *AmpTestSuite) SetupSuite() {
+	log.Println(">>>> Starting AMP Test Suite")
+}
 
-func init() {
-	environment.RegisterEnvironmentMetaDataFlags()
+func (suite *AmpTestSuite) TearDownSuite() {
+	suite.Result.Print()
+	log.Println(">>>> Finished AMP Test Suite")
+}
+
+func (suite *AmpTestSuite) TestAllInSuite() {
+	metadata = environment.GetEnvironmentMetaData()
 
 	ctx := context.Background()
 	var err error
@@ -72,18 +138,46 @@ func init() {
 	if err != nil {
 		fmt.Println("There was an error trying to load credentials: ", err)
 	}
+
+	for _, testRunner := range testRunners {
+		suite.AddToSuiteResult(testRunner.Run())
+	}
+	suite.Assert().Equal(status.SUCCESSFUL, suite.Result.GetStatus(), "Assume Role Test Suite Failed")
 }
 
-type AmpDestinationTestRunner struct {
+type AmpTestRunner struct {
 	test_runner.BaseTestRunner
+
+	source Source
+	config string
+	name   string
 }
 
-func (t AmpDestinationTestRunner) Validate() status.TestGroupResult {
-	metricsToFetch := t.GetMeasuredMetrics()
-	testResults := make([]status.TestResult, len(metricsToFetch))
+func (t AmpTestRunner) Validate() status.TestGroupResult {
+	// wait for agent to push some metrics
 	time.Sleep(30 * time.Second)
+
+	if t.source == SourcePrometheus {
+		return t.validatePrometheusMetrics()
+	} else if t.source == SourceHost {
+		return t.validateHostMetrics()
+	} else if t.source == SourceOtlp {
+		return t.validateOtlpMetrics()
+	}
+
+	return status.TestGroupResult{}
+}
+
+func (t *AmpTestRunner) validateHostMetrics() status.TestGroupResult {
+	// NOTE: dims must match aggregation_dimensions from agent config to fetch metrics.
+	// the idea is to fetch all metrics including non-aggregated metrics with matching dim set
+	// then validate if the returned list of metrics include metrics (non-aggregated) with append_dimensions as labels
+	dims := getDimensions()
+
+	metricsToFetch := t.getHostMetrics()
+	testResults := make([]status.TestResult, len(metricsToFetch))
 	for i, metricName := range metricsToFetch {
-		testResults[i] = t.validateMetric(metricName)
+		testResults[i] = t.validateStandardMetric(metricName, dims, true)
 	}
 
 	return status.TestGroupResult{
@@ -92,42 +186,73 @@ func (t AmpDestinationTestRunner) Validate() status.TestGroupResult {
 	}
 }
 
-func (t *AmpDestinationTestRunner) validateMetric(metricName string) status.TestResult {
-	env := environment.GetEnvironmentMetaData()
+func (t *AmpTestRunner) validatePrometheusMetrics() status.TestGroupResult {
+
+	dims := []types.Dimension{}
+
+	metricsToFetch := t.getPrometheusMetrics()
+	testResults := make([]status.TestResult, len(metricsToFetch))
+	for i, metricName := range metricsToFetch {
+		testResults[i] = t.validateStandardMetric(metricName, dims, false)
+	}
+
+	testResults = append(testResults, t.validateHistogramMetric("prometheus_test_histogram"))
+
+	return status.TestGroupResult{
+		Name:        t.GetTestName(),
+		TestResults: testResults,
+	}
+}
+
+func (t *AmpTestRunner) validateOtlpMetrics() status.TestGroupResult {
+
+	dims := []types.Dimension{}
+
+	metricsToFetch := t.getOtlpMetrics()
+	testResults := make([]status.TestResult, len(metricsToFetch))
+	for i, metricName := range metricsToFetch {
+		testResults[i] = t.validateStandardMetric(metricName, dims, false)
+	}
+
+	testResults = append(testResults, t.validateHistogramMetric("my_cumulative_histogram"))
+
+	return status.TestGroupResult{
+		Name:        t.GetTestName(),
+		TestResults: testResults,
+	}
+}
+
+func (t *AmpTestRunner) validateStandardMetric(metricName string, dims []types.Dimension, shouldHaveAppendDimensions bool) status.TestResult {
+	return t.validateMetric(ampQueryTemplate, metricName, dims, shouldHaveAppendDimensions)
+}
+
+func (t *AmpTestRunner) validateHistogramMetric(metricName string) status.TestResult {
+	return t.validateMetric(ampHistogramQueryTemplate, metricName, []types.Dimension{}, false)
+}
+
+func (t *AmpTestRunner) validateMetric(queryTemplate string, metricName string, dims []types.Dimension, shouldHaveAppendDimensions bool) status.TestResult {
 
 	testResult := status.TestResult{
 		Name:   metricName,
 		Status: status.FAILED,
 	}
 
-	// NOTE: dims must match aggregation_dimensions from agent config to fetch metrics.
-	// the idea is to fetch all metrics including non-aggregated metrics with matching dim set
-	// then validate if the returned list of metrics include metrics (non-aggregated) with append_dimensions as labels
-	dims := getDimensions()
-	if len(dims) == 0 {
-		return testResult
-	}
+	query := buildPrometheusQuery(queryTemplate, metricName, dims)
+	fmt.Printf("querying metrics for %s using the following query\n%s\n", metricName, query)
 
-	res, err := queryAMPMetrics(env.AmpWorkspaceId, buildPrometheusQuery(metricName, dims))
+	responseJSON, err := queryAMPMetrics(metadata.AmpWorkspaceId, query)
 	if err != nil {
 		fmt.Printf("failed to fetch metric values from AMP for %s: %s\n", metricName, err)
 		return testResult
 	}
-	var responseJson AMPResponse
-	err = json.Unmarshal(res, &responseJson)
-	if err != nil {
-		fmt.Printf("failed to unmarshal AMP response: %s\n", err)
-		return testResult
-	}
-
-	if len(responseJson.Data.Result) == 0 {
-		fmt.Printf("AMP metric values are missing for %s\n", metricName)
+	if len(responseJSON.Data.Result) == 0 {
+		fmt.Printf("failed because AMP metric values are missing for %s\n", metricName)
 		return testResult
 	}
 
 	foundAppendDimMetric := true
 	metricVals := []float64{}
-	for _, dataResult := range responseJson.Data.Result {
+	for _, dataResult := range responseJSON.Data.Result {
 		if len(dataResult.Value) < 1 {
 			continue
 		}
@@ -141,11 +266,24 @@ func (t *AmpDestinationTestRunner) validateMetric(metricName string) status.Test
 		}
 	}
 
-	// at least 2 metrics are expected with 1 set of aggregation_dimensions
-	// 1 non-aggregated + 1 aggregated minimum
-	if len(metricVals) < 2 || !foundAppendDimMetric {
-		fmt.Println("failed with less metric values than expected or missing append_dimensions")
-		return testResult
+	// AMP/Prometheus metrics do not have the appended dimensions
+	if shouldHaveAppendDimensions {
+		// at least 2 metrics are expected with 1 set of aggregation_dimensions
+		// 1 non-aggregated + 1 aggregated minimum
+		if len(metricVals) < 2 {
+			fmt.Println("failed with fewer metric values than expected")
+			return testResult
+		}
+
+		if !foundAppendDimMetric {
+			fmt.Println("failed with missing append_dimensions")
+			return testResult
+		}
+	} else {
+		if len(metricVals) < 1 {
+			fmt.Println("failed with fewer metric values than expected")
+			return testResult
+		}
 	}
 
 	if !metric.IsAllValuesGreaterThanOrEqualToExpectedValue(metricName, metricVals, 0) {
@@ -156,15 +294,26 @@ func (t *AmpDestinationTestRunner) validateMetric(metricName string) status.Test
 	return testResult
 }
 
-func (t AmpDestinationTestRunner) GetTestName() string {
-	return namespace
+func (t AmpTestRunner) GetTestName() string {
+	return t.name
 }
 
-func (t AmpDestinationTestRunner) GetAgentConfigFileName() string {
-	return "config.json"
+func (t AmpTestRunner) GetAgentConfigFileName() string {
+	return t.config
 }
 
-func (t AmpDestinationTestRunner) GetMeasuredMetrics() []string {
+func (t AmpTestRunner) GetMeasuredMetrics() []string {
+	if t.source == SourcePrometheus {
+		return t.getPrometheusMetrics()
+	} else if t.source == SourceHost {
+		return t.getHostMetrics()
+	} else if t.source == SourceOtlp {
+		return t.getOtlpMetrics()
+	}
+	return []string{}
+}
+
+func (t AmpTestRunner) getHostMetrics() []string {
 	return []string{
 		"CPU_USAGE_IDLE", "cpu_usage_nice", "cpu_usage_guest", "cpu_time_active", "cpu_usage_active",
 		"processes_blocked", "processes_dead", "processes_idle", "processes_paging", "processes_running",
@@ -173,40 +322,75 @@ func (t AmpDestinationTestRunner) GetMeasuredMetrics() []string {
 	}
 }
 
-func (t *AmpDestinationTestRunner) SetupBeforeAgentRun() error {
-	env := environment.GetEnvironmentMetaData()
+func (t AmpTestRunner) getPrometheusMetrics() []string {
+	return []string{
+		"prometheus_test_counter",
+		"prometheus_test_summary",
+	}
+}
+
+func (t AmpTestRunner) getOtlpMetrics() []string {
+	return []string{
+		"my_gauge",
+		"my_cumulative_counter",
+	}
+}
+
+func (t AmpTestRunner) SetupBeforeAgentRun() error {
 	err := t.BaseTestRunner.SetupBeforeAgentRun()
 	if err != nil {
 		return err
 	}
+
 	// replace AMP workspace ID placeholder with a testing workspace ID from metadata
 	agentConfigPath := filepath.Join("agent_configs", t.GetAgentConfigFileName())
 	ampCommands := []string{
-		"sed -ie 's/{workspace_id}/" + env.AmpWorkspaceId + "/g' " + agentConfigPath,
+		"sed -ie 's/{workspace_id}/" + metadata.AmpWorkspaceId + "/g' " + agentConfigPath,
 		// use below to add JMX metrics then update agent config & GetMeasuredMetrics()
 		//"nohup java -Dcom.sun.management.jmxremote -Dcom.sun.management.jmxremote.port=2030 -Dcom.sun.management.jmxremote.local.only=false -Dcom.sun.management.jmxremote.authenticate=false -Dcom.sun.management.jmxremote.ssl=false -Dcom.sun.management.jmxremote.rmi.port=2030  -Dcom.sun.management.jmxremote.host=0.0.0.0  -Djava.rmi.server.hostname=0.0.0.0 -Dserver.port=8090 -Dspring.application.admin.enabled=true -jar jars/spring-boot-web-starter-tomcat.jar > /tmp/spring-boot-web-starter-tomcat-jar.txt 2>&1 &",
 	}
 	err = common.RunCommands(ampCommands)
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to modify agent configuration: %w", err)
 	}
+
+	// Prometheus source has some special setup before the agent starts
+	if t.source == SourcePrometheus {
+		err = setupPrometheus()
+		if err != nil {
+			return fmt.Errorf("failed to setup prometheus: %w", err)
+		}
+	}
+
 	return t.SetUpConfig()
 }
 
-func TestAmp(t *testing.T) {
-	runner := test_runner.TestRunner{TestRunner: &AmpDestinationTestRunner{
-		test_runner.BaseTestRunner{},
-	}}
-	result := runner.Run()
-	if result.GetStatus() != status.SUCCESSFUL {
-		t.Fatal("AMP Destination test failed")
-		result.Print()
+func (t AmpTestRunner) SetupAfterAgentRun() error {
+
+	// OTLP source has some special setup after the agent starts
+	if t.source == SourceOtlp {
+		setupOtlp()
 	}
+
+	return nil
+}
+
+func setupPrometheus() error {
+	startPrometheusCommands := []string{
+		fmt.Sprintf("cat <<EOF | sudo tee /tmp/prometheus_config.yaml\n%s\nEOF", prometheusConfig),
+		fmt.Sprintf("cat <<EOF | sudo tee /tmp/metrics\n%s\nEOF", prometheusMetrics),
+		"sudo python3 -m http.server 8101 --directory /tmp &> /dev/null &",
+	}
+
+	return common.RunCommands(startPrometheusCommands)
+}
+
+func setupOtlp() error {
+	return common.RunAsyncCommand("resources/otlp_pusher.sh")
 }
 
 func getDimensions() []types.Dimension {
-	env := environment.GetEnvironmentMetaData()
-	factory := dimension.GetDimensionFactory(*env)
+	factory := dimension.GetDimensionFactory(*metadata)
 	dims, failed := factory.GetDimensions([]dimension.Instruction{
 		{
 			Key:   "InstanceId",
@@ -225,7 +409,7 @@ func getDimensions() []types.Dimension {
 	return dims
 }
 
-func buildPrometheusQuery(metricName string, dims []types.Dimension) string {
+func buildPrometheusQuery(template string, metricName string, dims []types.Dimension) string {
 	dimsStr := ""
 	for _, dim := range dims {
 		dimsStr = fmt.Sprintf("%s%s=\"%s\", ", dimsStr, *dim.Name, *dim.Value)
@@ -233,34 +417,46 @@ func buildPrometheusQuery(metricName string, dims []types.Dimension) string {
 	if len(dimsStr) > 0 {
 		dimsStr = dimsStr[:len(dimsStr)-1]
 	}
-	return fmt.Sprintf(ampQueryTemplate, metricName, "{"+dimsStr+"}")
+	return fmt.Sprintf(template, metricName, "{"+dimsStr+"}")
 }
 
-func queryAMPMetrics(wsId string, q string) ([]byte, error) {
-	url := fmt.Sprintf("https://aps-workspaces.%s.amazonaws.com/workspaces/%s/api/v1/query?query=%s", awsConfig.Region, wsId, q)
+func queryAMPMetrics(wsID string, q string) (AMPResponse, error) {
+	url := fmt.Sprintf("https://aps-workspaces.%s.amazonaws.com/workspaces/%s/api/v1/query?query=%s", awsConfig.Region, wsID, q)
 	req, err := http.NewRequest(http.MethodPost, url, nil)
 	if err != nil {
-		return nil, err
+		return AMPResponse{}, fmt.Errorf("failed to create AMP request: %w", err)
 	}
 
 	signer := sigv4.NewSigner()
 	err = signer.SignHTTP(context.Background(), awsCreds, req, hex.EncodeToString(sha256.New().Sum(nil)), "aps", awsConfig.Region, time.Now().UTC())
 	if err != nil {
-		return nil, err
+		return AMPResponse{}, fmt.Errorf("failed to sign AMP request: %w", err)
 	}
 
 	res, err := http.DefaultClient.Do(req)
 	if err != nil {
-		return nil, err
+		return AMPResponse{}, fmt.Errorf("failed to send AMP request: %w", err)
 	}
-	return io.ReadAll(res.Body)
+
+	bytes, err := io.ReadAll(res.Body)
+	if err != nil {
+		return AMPResponse{}, fmt.Errorf("failed to read AMP response: %w", err)
+	}
+
+	var responseJSON AMPResponse
+	err = json.Unmarshal(bytes, &responseJSON)
+	if err != nil {
+		return AMPResponse{}, fmt.Errorf("failed to unmarshal AMP response: %w", err)
+	}
+
+	return responseJSON, nil
 }
 
 func matchDimensions(labels map[string]interface{}) bool {
-	if len(append_dims) > len(labels) {
+	if len(appendDims) > len(labels) {
 		return false
 	}
-	for k, v := range append_dims {
+	for k, v := range appendDims {
 		if lv, found := labels[k]; !found || lv != v {
 			return false
 		}
@@ -268,4 +464,4 @@ func matchDimensions(labels map[string]interface{}) bool {
 	return true
 }
 
-var _ test_runner.ITestRunner = (*AmpDestinationTestRunner)(nil)
+var _ test_runner.ITestRunner = (*AmpTestRunner)(nil)
