+++ test/amp/destination/resources/otlp_metrics.json
