+++ test/amp/destination/resources/prometheus.yaml
